requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
